% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasKey.R
\name{hasKey}
\alias{hasKey}
\title{hasKey}
\usage{
hasKey(x, ...)
}
\arguments{
\item{x}{listArray object}

\item{...}{index to specify elements to check}
}
\value{
logical
}
\description{
Check is a specific index exists
}
\examples{
l <- listArray()
l[1] <- 1
hasKey(l, 1)
l[2,3] <- "test"
hasKey(l, 2, 3)
l[2:3] <- "vector"
hasKey(l, 2:3)
l['iris'] <- iris
hasKey(l, iris) # FALSE
l[mean] <- mean
hasKey(l, mean)
# if you have not stored NULL objects in your listArray
is.null(l[mean])
is.null(l['iris'])
}
