% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\title{key}
\usage{
key(...)
}
\arguments{
\item{...}{R objects}
}
\value{
a unique character key
}
\description{
Creates a character key from arbitray R objects.
}
\examples{
key(1)
key(2,3)
key(1:3)
key(mean)
key('test')
key(letters[1:5])
key(list(1))
}
