% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys}
\alias{keys}
\title{Returns a list of indices}
\usage{
keys(x)
}
\arguments{
\item{x}{listArray obkect}
}
\value{
a character vector with the retranslated indices
}
\description{
Returns a list of indices
}
\examples{
l <- listArray(1:5)
k <- keys(l)
k
# access object in listArray
pos <- which(k=='3')
k[[pos]]
}
